+++ packages/datadog-plugin-aws-sdk/src/index.js
@@ -45,18 +45,7 @@ function createWrapRequest (tracer, config) {
       return tracer.scope().activate(span, () => {
         let boundCb
         if (typeof cb === 'function') {
-          boundCb = function wrappedCb (err, resp) {
-            const maybeChildOf = awsHelpers.responseExtract(serviceName, request, resp, tracer)
-            if (maybeChildOf) {
-              const options = {
-                childOf: maybeChildOf,
-                tags: Object.assign({}, tags, { [Tags.SPAN_KIND]: 'server' })
-              }
-              boundCb = tracer.wrap('aws.response', options, cb).call(this, err, resp)
-            } else {
-              boundCb = tracer.scope().bind(cb, childOf).call(this, err, resp)
-            }
-          }
+          boundCb = awsHelpers.wrapCb(cb, serviceName, tags, request, tracer, childOf)
         } else {
           boundCb = cb
         }
