+++ packages/datadog-plugin-aws-sdk/test/fixtures/sqs.js
@@ -5,7 +5,6 @@ const sqs = {}
 sqs['create'] = {
   QueueName: 'SQS_QUEUE_NAME',
   Attributes: {
-    // 'DelaySeconds': '60',
     'MessageRetentionPeriod': '86400'
   }
 }
